Compatibility Updates
Included the player_crisis_chosen flag in my AP overrides - and Condans AP itself.  This is for Modjam 2025 compatibility.

Moved the AP overrides that care specifically just about not being a player crisis into their own file so they are easier to read.

Removed a vanilla override for turning a planet into the tomb world by writing my own parallel on_action event to trigger and sort that in the background.